removed help screen.  refactored some contact_util stuff.  refactored app drawer.
