+++ lib/services/background_service.dart
@@ -6,7 +6,7 @@ import 'package:recall/repositories/contact_repository.dart';
 import 'package:recall/repositories/usersettings_repository.dart';
 import 'package:recall/services/notification_helper.dart';
 import 'package:recall/services/notification_service.dart';
-import 'package:recall/utils/last_contacted_utils.dart';
+import 'package:recall/utils/contact_utils.dart';
 import 'package:workmanager/workmanager.dart';
 import 'package:recall/objectbox.g.dart'; // Import generated ObjectBox code
 import 'package:path_provider/path_provider.dart'; // Import path_provider
